Reversed some disabled definitions regarding CO2eq mass quantities and units, and corrected the units definitions
I've slept a night about it and have come to the conclusion that g CO2eq and the like (t CO2eq, kg CO2eq) _are_ mass units, but their definitions are not fixed like normally is the case with units (and is the _idea_ with units). One does also encounter this with currencies, which are time-dependent. In the case of the CO2eq mass units they are dependent on the measured phenomenon itself, as Jan Martin pointed out. Presently we can not indicate such formally in OM yet, so the definitions of these units in grams remain disabled, but the other definitions/productions that I have disabled yesterday, can be reversed, which I have now done.

I wonder by the way why we originally have defined tonneOfCarbonDioxideEquivalent in (normal) tonnes. It seems more logical to define it as 1000 kilogramOfCarbonDioxideEquivalent, which on its turn has already correctly been defined as a prefixed unit in terms of kilo and gramOfCarbonDioxideEquivalent. So, I have also changed this, accordingly.

By the way (2), I think we should define gramOfCarbonDioxideEquivalent as 0.001 kilogramOfCarbonDioxideEquivalent, just like gram is officially defined in terms of kilogram in the SI (I admit, this is strange). Until yesterday, gramOfCarbonDioxideEquivalent was defined as 1 gram, and since yesterday - as described above - this definition has been disabled. Now I have changed this accordingly.